Avance nuevas conexiones

+++ API/Controllers/Profesor/obtenerSoliLab.cs
@@ -40,14 +40,14 @@ public class ObtSolisLabController : Controller
                                     CorreoSoli = reader.GetString(0),
                                     FechaSoli = reader.GetDateTime(1),
                                     HoraSoli = reader.GetTimeSpan(2),
-                                    Carnet = reader.GetInt32(3),
+                                    Carnet = reader.GetDecimal(3),
                                     PNombre = reader.GetString(4),
                                     SNombre = reader.GetString(5),
                                     PApellido = reader.GetString(6),
                                     SApellido = reader.GetString(7),
                                     CantHoras = reader.GetDecimal(8),
                                     LabNombre = reader.GetString(9),
-                                    UserCed = reader.GetInt32(10)
+                                    UserCed = reader.GetDecimal(10)
                                 });
                             }
 
@@ -73,12 +73,12 @@ public class SolicitudLabResponse
     public string CorreoSoli { get; set; }
     public DateTime FechaSoli { get; set; }
     public TimeSpan HoraSoli { get; set; }
-    public int Carnet { get; set; }
+    public decimal Carnet { get; set; }
     public string PNombre { get; set; }
     public string SNombre { get; set; }
     public string PApellido { get; set; }
     public string SApellido { get; set; }
     public decimal CantHoras { get; set; }
     public string LabNombre { get; set; }
-    public int UserCed { get; set; }
+    public decimal UserCed { get; set; }
 }
